requirements.txt
aiogram==3.15.0
PyMuPDF==1.25.3
pypdf==5.4.0
Pillow==11.1.0
img2pdf==0.5.1
aiofiles==24.1.0
aiosqlite==0.21.0
pydantic-settings==2.7.1
python-dotenv==1.1.0
python-magic==0.4.27
